+++ ircd/m_svsmode.c
@@ -126,3 +126,37 @@ int ms_svsmode(struct Client* cptr, struct Client* sptr, int parc, char* parv[])
   return 0;
 }
 
+
+/** Handle SVSMODE from a local operator.
+ * Requires PRIV_NETADMIN privilege.
+ * parv[1] = Target nickname
+ * parv[2] = Mode changes
+ */
+int mo_svsmode(struct Client* cptr, struct Client* sptr, int parc, char* parv[])
+{
+  struct Client *acptr;
+
+  if (!HasPriv(sptr, PRIV_NETADMIN))
+    return send_reply(sptr, ERR_NOPRIVILEGES);
+
+  if (parc < 3)
+    return need_more_params(sptr, "SVSMODE");
+
+  if (!(acptr = FindUser(parv[1])))
+    return send_reply(sptr, ERR_NOSUCHNICK, parv[1]);
+
+  sendto_opmask_butone(0, SNO_OLDSNO, "%C used SVSMODE on %C: %s",
+                        sptr, acptr, parv[2]);
+
+  if (MyUser(acptr)) {
+    char *args[4];
+    args[0] = parv[0];
+    args[1] = (char *)NumNick(acptr);
+    args[2] = parv[2];
+    args[3] = NULL;
+    return ms_svsmode(cptr, &me, 3, args);
+  }
+
+  sendcmdto_serv_butone(&me, CMD_SVSMODE, cptr, "%C %s", acptr, parv[2]);
+  return 0;
+}